requirements.txt
Django==2.0.8

django-webpack-loader==0.5.0

pytz==2017.3

djangorestframework==3.9.0